Move show-policy report generation to a service object
